GH-45279: [C++][Compute] Move all Grouper tests to grouper_test.cc
