+++ cpp/src/arrow/acero/aggregate_node.h
@@ -24,6 +24,7 @@
 
 #include "arrow/acero/visibility.h"
 #include "arrow/compute/api_aggregate.h"
+#include "arrow/compute/test_util_internal.h"
 #include "arrow/compute/type_fwd.h"
 #include "arrow/result.h"
 #include "arrow/type_fwd.h"
