+++ cpp/src/arrow/acero/asof_join_node_test.cc
@@ -41,8 +41,9 @@
 #include "arrow/acero/util.h"
 #include "arrow/api.h"
 #include "arrow/compute/api_scalar.h"
-#include "arrow/compute/kernels/test_util.h"
+#include "arrow/compute/cast.h"
 #include "arrow/compute/row/row_encoder_internal.h"
+#include "arrow/compute/test_util_internal.h"
 #include "arrow/testing/gtest_util.h"
 #include "arrow/testing/matchers.h"
 #include "arrow/testing/random.h"
