+++ cpp/src/arrow/compute/CMakeLists.txt
@@ -28,6 +28,14 @@ endif()
 # Unit tests
 #
 
+# Define arrow_compute_testing object library for common test files
+if(ARROW_TESTING)
+  add_library(arrow_compute_testing OBJECT test_util_internal.cc)
+  # Even though this is still just an object library we still need to "link" our
+  # dependencies so that include paths are configured correctly
+  target_link_libraries(arrow_compute_testing PUBLIC ${ARROW_GTEST_GMOCK})
+endif()
+
 set(ARROW_COMPUTE_TEST_PREFIX "arrow-compute")
 set(ARROW_COMPUTE_TEST_LABELS "arrow-compute-tests")
 set(ARROW_COMPUTE_TEST_ARGS PREFIX ${ARROW_COMPUTE_TEST_PREFIX} LABELS
@@ -87,9 +95,16 @@ add_arrow_test(internals_test
                function_test.cc
                exec_test.cc
                kernel_test.cc
-               registry_test.cc)
+               registry_test.cc
+               EXTRA_LINK_LIBS
+               arrow_compute_testing)
 
-add_arrow_compute_test(expression_test SOURCES expression_test.cc)
+add_arrow_compute_test(expression_test
+                       SOURCES
+                       expression_test.cc
+                       EXTRA_LINK_LIBS
+                       arrow_compute_testing)
+
 add_arrow_compute_test(row_test
                        SOURCES
                        key_hash_test.cc
@@ -98,7 +113,9 @@ add_arrow_compute_test(row_test
                        row/grouper_test.cc
                        row/row_encoder_internal_test.cc
                        row/row_test.cc
-                       util_internal_test.cc)
+                       util_internal_test.cc
+                       EXTRA_LINK_LIBS
+                       arrow_compute_testing)
 
 add_arrow_benchmark(function_benchmark PREFIX "arrow-compute")
