+++ cpp/src/arrow/compute/kernels/scalar_boolean_test.cc
@@ -24,7 +24,7 @@
 
 #include "arrow/chunked_array.h"
 #include "arrow/compute/api_scalar.h"
-#include "arrow/compute/kernels/test_util.h"
+#include "arrow/compute/kernels/test_util_internal.h"
 #include "arrow/testing/gtest_util.h"
 #include "arrow/util/checked_cast.h"
