+++ cpp/src/arrow/compute/kernels/vector_cumulative_ops_test.cc
@@ -31,7 +31,7 @@
 
 #include "arrow/array/builder_primitive.h"
 #include "arrow/compute/api.h"
-#include "arrow/compute/kernels/test_util.h"
+#include "arrow/compute/kernels/test_util_internal.h"
 #include "arrow/type_fwd.h"
 
 namespace arrow {
